requirements.txt
fdb==2.0.2
future==1.0.0
numpy==2.2.1
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.17.0
tzdata==2024.2
